# openapi.json
{
  "components": {
    "schemas": {
...
      "Envelope_NotificationResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/NotificationResponse"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[NotificationResponse]",
        "type": "object"
      },
...
      "Envelope_ReadAllResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ReadAllResponse"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[ReadAllResponse]",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "도메인 에러 코드. 분기는 message 가 아니라 이 값으로 한다",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "NotificationPage": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/NotificationPageMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "NotificationPage",
        "type": "object"
      },
      "NotificationPageMeta": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "다음 페이지 커서. null 이면 마지막 페이지다. 다음 요청의 `cursor` 에 넣는다",
            "title": "Next Cursor"
          },
          "unread_count": {
            "description": "필터와 무관한 내 안읽은 알림 전체 개수",
            "title": "Unread Count",
            "type": "integer"
          }
        },
        "required": [
          "next_cursor",
          "unread_count"
        ],
        "title": "NotificationPageMeta",
        "type": "object"
      },
      "NotificationRead": {
        "properties": {
          "read": {
            "description": "true 면 읽음, false 면 다시 안읽음으로 되돌린다",
            "title": "Read",
            "type": "boolean"
          }
        },
        "required": [
          "read"
        ],
        "title": "NotificationRead",
        "type": "object"
      },
      "NotificationResponse": {
        "properties": {
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "channels_sent": {
            "items": {
              "type": "string"
            },
            "title": "Channels Sent",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "notification_id": {
            "format": "uuid",
            "title": "Notification Id",
            "type": "string"
          },
          "read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "related_ref": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RelatedRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "notification_id",
          "user_id",
          "type",
          "title",
          "body",
          "related_ref",
          "channels_sent",
          "read_at",
          "created_at"
        ],
        "title": "NotificationResponse",
        "type": "object"
      },
...
      "ReadAllResponse": {
        "properties": {
          "updated_count": {
            "title": "Updated Count",
            "type": "integer"
          }
        },
        "required": [
          "updated_count"
        ],
        "title": "ReadAllResponse",
        "type": "object"
      },
...
      "RelatedRef": {
        "description": "명세 §2.1 의 `related_ref`. DB 는 resource_type · resource_id 두 컬럼으로 갖는다.",
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "id"
        ],
        "title": "RelatedRef",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "성공 응답은 `{data, meta}`, 실패는 `{error: {code, message, details}}` 다. 분기는 HTTP 상태가 아니라 `error.code` 로 한다 — 같은 상태 코드에 코드가 여럿이다.\n\n입력 검증 실패는 전부 **400 `INVALID_INPUT`** 이며 `details.fields` 에 위치가 온다. 인증은 `Authorization: Bearer <access_token>`. 목록은 커서 페이지네이션이다: `meta.next_cursor` 를 다음 요청의 `cursor` 로 넘기고, null 이면 끝이다.",
    "title": "Prolog (StoryForge) API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/notifications": {
      "get": {
        "description": "최근 알림부터. 커서 페이지네이션이다.\n\n`meta.unread_count` 는 필터와 무관하게 내 안읽은 알림 전체 개수다(배지용).",
        "operationId": "list_notifications_v1_notifications_get",
        "parameters": [
          {
            "description": "페이지 크기. 기본 20 · 최대 100",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "페이지 크기. 기본 20 · 최대 100",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "이전 응답의 `meta.next_cursor`. 첫 페이지는 비운다",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "이전 응답의 `meta.next_cursor`. 첫 페이지는 비운다",
              "title": "Cursor"
            }
          },
          {
            "description": "true 면 안읽은 알림만",
            "in": "query",
            "name": "unread_only",
            "required": false,
            "schema": {
              "default": false,
              "description": "true 면 안읽은 알림만",
              "title": "Unread Only",
              "type": "boolean"
            }
          },
          {
            "description": "알림 종류로 거른다. 예: team_invite",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "알림 종류로 거른다. 예: team_invite",
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "내 알림 목록",
        "tags": [
          "NOTI"
        ]
      }
    },
    "/v1/notifications/read-all": {
      "patch": {
        "description": "`updated_count` 는 이번 호출로 새로 읽음이 된 개수다.",
        "operationId": "mark_all_read_v1_notifications_read_all_patch",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ReadAllResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "내 알림을 모두 읽음으로 표시한다",
        "tags": [
          "NOTI"
        ]
      }
    },
    "/v1/notifications/{notificationId}": {
      "delete": {
        "description": "내 알림만 지울 수 있다.",
        "operationId": "delete_notification_v1_notifications__notificationId__delete",
        "parameters": [
          {
            "in": "path",
            "name": "notificationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notificationid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "NOTIFICATION_NOT_FOUND": {
                    "summary": "알림을 찾을 수 없습니다",
                    "value": {
                      "error": {
                        "code": "NOTIFICATION_NOT_FOUND",
                        "details": {},
                        "message": "알림을 찾을 수 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `NOTIFICATION_NOT_FOUND` — 알림을 찾을 수 없습니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "알림을 삭제한다",
        "tags": [
          "NOTI"
        ]
      },
      "get": {
        "description": "내 알림만 볼 수 있다. 남의 알림은 존재하지 않는 것으로 답한다.",
        "operationId": "get_notification_v1_notifications__notificationId__get",
        "parameters": [
          {
            "in": "path",
            "name": "notificationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notificationid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_NotificationResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "NOTIFICATION_NOT_FOUND": {
                    "summary": "알림을 찾을 수 없습니다",
                    "value": {
                      "error": {
                        "code": "NOTIFICATION_NOT_FOUND",
                        "details": {},
                        "message": "알림을 찾을 수 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `NOTIFICATION_NOT_FOUND` — 알림을 찾을 수 없습니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "알림을 조회한다",
        "tags": [
          "NOTI"
        ]
      },
      "patch": {
        "description": "`read = false` 로 다시 안읽음으로 되돌릴 수 있다.",
        "operationId": "set_read_v1_notifications__notificationId__patch",
        "parameters": [
          {
            "in": "path",
            "name": "notificationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notificationid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationRead"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_NotificationResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_INPUT": {
                    "summary": "요청 값이 올바르지 않습니다",
                    "value": {
                      "error": {
                        "code": "INVALID_INPUT",
                        "details": {},
                        "message": "요청 값이 올바르지 않습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `INVALID_INPUT` — 요청 값이 올바르지 않습니다"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "인증이 필요합니다",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "details": {},
                        "message": "인증이 필요합니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `UNAUTHORIZED` — 인증이 필요합니다"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "NOTIFICATION_NOT_FOUND": {
                    "summary": "알림을 찾을 수 없습니다",
                    "value": {
                      "error": {
                        "code": "NOTIFICATION_NOT_FOUND",
                        "details": {},
                        "message": "알림을 찾을 수 없습니다"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "- `NOTIFICATION_NOT_FOUND` — 알림을 찾을 수 없습니다"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "알림의 읽음 상태를 바꾼다",
        "tags": [
          "NOTI"
        ]
      }
    },
...
  }
}